Remove initial migration file
This commit involves the deletion of the initial migration file for SPCaemucals.Data. The removal is part of database schema changes as the existing migrations do not match with the current structure of the application database. The corresponding schema and related tables creation codes have been removed as well.

add